missing limitation in AppServiceFileAuditLogs
In Japanese troubleshoot-diagnostic-logs.md, there is missing limitation in AppServiceFileAuditLogs.
In the following English version, there is the limitation of  'only available for Premium tier and above' at the end of its explanation cell in AppServiceFileAuditLogs record.

https://docs.microsoft.com/en-us/azure/app-service/troubleshoot-diagnostic-logs#supported-log-types

But in Japanese version, the above limitation is not documented. This PR try to add the limitation in Japanese document not to lead misunderstanding of our customers.